refactored getAbility and helperAbs -> moved helper abilities to appropriate classes.
